login/signup check

+++ ContestAi/ContestAdmin/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render,redirect
 from .forms import *
+from django.contrib.auth.models import User
 # Create your views here.
 
 def index(request):
@@ -10,12 +11,14 @@ def addStatus(request):
         staticF = StaticForm(request.POST, request.FILES)
         if staticF.is_valid():
             staticF.save()
-
-    staticF = StaticForm()
-    context = {
-        'formStatus': staticF
-    }
-    return render(request, 'Status/add.html', context=context)
+    if request.user.is_superuser:
+        staticF = StaticForm()
+        context = {
+            'formStatus': staticF
+        }
+        return render(request, 'Status/add.html', context=context)
+    else:
+        return redirect('login')
 
 def updateStatus(request, pk):
     istatus = Status.objects.get(pk=pk)
@@ -23,30 +26,40 @@ def updateStatus(request, pk):
         staticF = StaticForm(request.POST, request.FILES,instance=istatus)
         if staticF.is_valid():
             staticF.save()
-            
-    staticF = StaticForm(instance=istatus)
-    context = {
-        'formStatus': staticF
-    }
-    print(context)
-    return render(request, 'Status/add.html', context=context)
+    if request.user.is_superuser:      
+        staticF = StaticForm(instance=istatus)
+        context = {
+            'formStatus': staticF
+        }
+        print(context)
+        return render(request, 'Status/add.html', context=context)
+    else:
+        return redirect('login')
 
 def listStatus(request):
-    lstatus = Status.objects.all()
-    context = {
-        'lstatus': lstatus
-    }
-    return render(request, 'Status/list.html', context=context)
+    if request.user.is_superuser:
+        lstatus = Status.objects.all()
+        context = {
+            'lstatus': lstatus
+        }
+        return render(request, 'Status/list.html', context=context)
+    else:
+        return redirect('login')
 
 def detailStatus(request, pk):
-    istatus = Status.objects.get(pk=pk)
-    context = {
-        'istatus':istatus
-    }
-    return render(request, 'Status/detail.html', context=context)
+    if request.user.is_superuser:
+        istatus = Status.objects.get(pk=pk)
+        context = {
+            'istatus':istatus
+        }
+        return render(request, 'Status/detail.html', context=context)
+    else:
+        return redirect('login')
 
 def deleteStatus(request, pk):
-    dstatus = Status.objects.get(pk=pk)
-    dstatus.delete()
-    return listStatus(request)
-    
+    if request.user.is_superuser:
+        dstatus = Status.objects.get(pk=pk)
+        dstatus.delete()
+        return listStatus(request)
+    else:
+        return redirect('login')
